Relax testing of graph handing in Queryable specs
`Queryable` shared examples failed for `Graph`s with a non-default graph
name. They required the graph to retain the query pattern's
`graph_name`, which is undesirable behavior. Instead `Graph` replaces
the provided `graph_name` with its own, when querying, mimicking its
behavior on write operations.

This relaxes the specs to allow for the special case of a queryable
scoped in this way.